app/requirements.txt
alembic>=1.13.1
asyncpg>=0.29.0
asyncio>=3.4.3
bcrypt>=4.1.2
decorator>=5.1.1
fastapi>0.110.0
fastapi-jwt-auth>=0.5.0
python-jose>=1.0.0
psycopg2>=2.9.9
passlib>=1.7.4
pydantic>=2.6.4
pydantic_core>=2.16.3
pydantic_settings>=2.2.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-multipart>=0.0.9
SQLAlchemy>=2.0.28
SQLAlchemy-Utils>=0.41.2
starlette>=0.36.3
typing>=3.7.4.3
typing-extensions>=4.10.0
uvicorn>=0.28.1
